requirements.txt
astroid==3.0.2
coverage==7.4.1
dill==0.3.8
exceptiongroup==1.2.0
Faker==23.1.0
iniconfig==2.0.0
install==1.3.5
isort==5.13.2
mccabe==0.7.0
packaging==23.2
platformdirs==4.1.0
pluggy==1.4.0
pylint==3.0.3
pylint-pytest==1.1.7
pytest==8.0.0
pytest-cov==4.1.0
pytest-pylint==0.21.0
python-dateutil==2.8.2
python-dotenv==1.0.1
six==1.16.0
tomli==2.0.1
tomlkit==0.12.3
typing_extensions==4.9.0